Casts to remove Windows warning

+++ src/error/error.c
@@ -18,7 +18,7 @@ int JEL_error_push(struct JEL_Error error)
 {
   if (JEL_context_current->error_stack->errors_num < JEL_ERROR_STACK_ERRORS_MAX) {
     if (JEL_context_current->error_stack->errors_allocated <= JEL_context_current->error_stack->errors_num) {
-      if (JEL_error_stack_errors_allocate_p(JEL_context_current->error_stack, JEL_context_current->error_stack->errors_allocated * 1.618)) {
+      if (JEL_error_stack_errors_allocate_p(JEL_context_current->error_stack, (uint32_t) (JEL_context_current->error_stack->errors_allocated * 1.618))) {
         struct JEL_Error e = {"Cannot allocate more errors in JEL_error_push", JEL_ERROR_ALLOCATE};
         JEL_error_push(e);
         return JEL_ERROR_ALLOCATE;
